rename ds packages
