+++ pkg/engine/datasource/graphql_datasource/graphql_datasource_test.go
@@ -1,4 +1,4 @@
-package graphqldatasourcev2
+package graphql_datasource
 
 import (
 	"encoding/json"
