+++ pkg/engine/datasource/rest_datasource/httpjson_datasource_test.go
@@ -1,4 +1,4 @@
-package httpjsondatasource
+package rest_datasource
 
 import (
 	"context"
